day-3/lab-4/requirements.txt
langchain
langchain-community
langchain-openai
langgraph
chromadb
streamlit
duckduckgo-search
ddgs
pandas
python-dotenv
langsmith
